+++ jsolex-core/src/main/java/me/champeau/a4j/jsolex/processing/expr/impl/GeometryCorrection.java
@@ -74,7 +74,6 @@ public class GeometryCorrection extends AbstractFunctionImpl {
                 image.findMetadata(Ellipse.class).orElse(null),
                 null,
                 null,
-                null,
                 0,
                 getFromContext(ProcessParams.class).orElse(ProcessParams.loadDefaults()),
                 getFromContext(ImageEmitter.class).orElse(null),